refactor(react-scheduling): move the availability helpers out of core
`getEffectiveAvailability` and `setScheduleAvailability`, and the private
functions they are built from, move to `react-scheduling/src/availability.ts`.
Core keeps the SchedulingParameters plumbing underneath them, which is untyped
by design: a parameter is any sub-extension, and the server reads it that way.
The typed `HealthcareServiceAvailableTime` layer is tied to the editing use
case, and this package is alpha, so it can change shape freely here.

Tests split along the same line. The four cases that only reached core's own
behavior through `setScheduleAvailability` -- version-suffixed service
references, creating the SchedulingParameters container, not leaving a stale
value on a second matching extension, and scoped clearing -- are rewritten
against `setScheduleParameter` with a plain `bufferBefore` parameter, so that
behavior stays tested where it lives.
