+++ examples/medplum-provider/src/components/ReleaseStageBanner.module.css
@@ -9,7 +9,6 @@
   font-weight: 600;
 }
 
-/* Alpha keeps the more eye-catching violet; blue above is the beta and default hue. */
 .banner.alpha {
   background-color: light-dark(var(--mantine-color-violet-1), var(--mantine-color-violet-9));
 }
