+++ packages/docs/docs/scheduling/defining-availability.md
@@ -308,7 +308,7 @@ In this example:
 
 Rather than hand-authoring the [`availability` extension](#availability-extension), the [`@medplum/react-scheduling`](https://www.npmjs.com/package/@medplum/react-scheduling) library provides a `ScheduleAvailabilityEditor` component. It edits a Schedule's weekly `availability` override for a given visit service type, or, with the `schedule` prop omitted, the [service-level default](#service-level-availability) hours themselves. It implements the [override behavior](#override-behavior) described above through a single switch, and is used in the [Medplum Provider](https://github.com/medplum/medplum/tree/main/examples/medplum-provider) example app.
 
-`@medplum/core` exports the framework-agnostic helpers the component is built on, for reading and writing the `availability` override without the UI (for example, in a bot or a custom editor).
+The helpers the component reads and writes the override through, `getEffectiveAvailability` and `setScheduleAvailability`, are exported from the same library. They are plain functions over FHIR resources, so a custom editor can use them without the component. `@medplum/core` holds the generic scheduling parameter helpers underneath them — `getScheduleParameters`, `setScheduleParameter`, and `clearScheduleParameter` — which read and write any parameter, `availability` included, as extensions.
 
 See the [`ScheduleAvailabilityEditor` docs in Storybook](https://storybook.medplum.com/?path=/docs/medplum-scheduleavailabilityeditor--docs) for interactive examples, the behavior in detail, and the full component and utility API.
