+++ packages/react-scheduling/src/index.ts
@@ -6,3 +6,6 @@ export type * from './types';
 // Export all components
 export * from './Calendar/Calendar';
 export * from './ScheduleAvailabilityEditor/ScheduleAvailabilityEditor';
+
+// Helpers the components are built on, usable without them
+export * from './availability';
